refactor: move near Nano app JS bindings to package
