+++ libs/ledger-live-common/package.json
@@ -134,6 +134,7 @@
     "@ledgerhq/hw-app-btc": "workspace:^",
     "@ledgerhq/hw-app-cosmos": "workspace:^",
     "@ledgerhq/hw-app-eth": "workspace:^",
+    "@ledgerhq/hw-app-near": "workspace:^",
     "@ledgerhq/hw-app-polkadot": "workspace:^",
     "@ledgerhq/hw-app-solana": "workspace:^",
     "@ledgerhq/hw-app-str": "workspace:^",
