+++ libs/ledgerjs/packages/hw-app-near/CHANGELOG.md
@@ -1,0 +1,31 @@
+# @ledgerhq/hw-app-near
+
+## 6.27.3
+
+### Patch Changes
+
+- Updated dependencies [[`ecfdd1ebd8`](https://github.com/LedgerHQ/ledger-live/commit/ecfdd1ebd8cc7c4b5bc6315316ce662bb6241311)]:
+  - @ledgerhq/hw-transport@6.27.3
+
+## 6.27.3-next.0
+
+### Patch Changes
+
+- Updated dependencies [[`ecfdd1ebd`](https://github.com/LedgerHQ/ledger-live/commit/ecfdd1ebd8cc7c4b5bc6315316ce662bb6241311)]:
+  - @ledgerhq/hw-transport@6.27.3-next.0
+
+## 6.27.2
+
+### Patch Changes
+
+- Updated dependencies [[`203b927b4`](https://github.com/LedgerHQ/ledger-live/commit/203b927b4e5bca3402c85a88c536d519adb18c5f)]:
+  - @ledgerhq/errors@6.10.1
+  - @ledgerhq/hw-transport@6.27.2
+
+## 6.27.2-next.0
+
+### Patch Changes
+
+- Updated dependencies [[`203b927b4`](https://github.com/LedgerHQ/ledger-live/commit/203b927b4e5bca3402c85a88c536d519adb18c5f)]:
+  - @ledgerhq/errors@6.10.1-next.0
+  - @ledgerhq/hw-transport@6.27.2-next.0
