+++ libs/ledgerjs/packages/hw-app-near/jest.config.ts
@@ -1,0 +1,6 @@
+import baseConfig from "../../jest.config";
+
+export default {
+  ...baseConfig,
+  rootDir: __dirname,
+};
